make it work

+++ src/MPC.h
@@ -7,7 +7,7 @@
 using namespace std;
 
 // latency in ms
-const double latency_ms=100;
+const int latency_ms=100;
 // Already tuned by given project -> no tuning anymore
 // Lf was tuned until the the radius formed by the simulating the model
 // presented in the classroom matched the previous radius.
